Add a shim for entity_language to title
entity_language used to use the 'label callback' to determine correct
label value for a given node. entity_language was removed in Backdrop but
there is already a title_entity_language function in the title module
which returns a different value than this because it uses the callback.
So I've named so that it's clear this is a shim and replaced calls to
entity_language in title with the shim function. The calls to
title_language I left alone.